Updated admin functionality
Changed user collection to have points, added an edit session for admins thats viewable from the admin tab (need to ask aaron how to make sure that the old data is showed in form), and added view session to the admin page.
todo:  add buttons to viewsessionadmin that will 1: send admin to edit page and 2: give admin the option to delete any session.